Initialize Chatroom Page and Fix Sidebar Display Issues (#7)
This PR includes the chatroom feature of Grickly using Socket. It is added to
the sidebar and allows users who are logged in to communicate with each other.
The user can see all his/her chatrooms and switch among them. It also displays
all members in a chatroom so the user can see group members' profile. The page
is set to show the first chatroom that the user is in when the user refresh the page.

The sidebar display issues of not displaying and overlaying the page components
are also fixed. The reason that the sidebar is not showing is `onLeftIconButtonTouchTap`
in `AppBar` is not accessible for some.

Co-authored-by: Sicong Liu <sicong.liu.simon@outlook.com>