Update Theme Color and Layout of All Pages (#13)
This PR updates the theme color and UI layout of the site to make it looks more
professional. The theme color was changed from a dark theme to a lighter theme.
The layout of all pages was changed to fit the new theme. The location of the register
and reset password page was also moved for better user experience.

Co-authored-by: HengningZhang <hz1704@nyu.edu>